Structure MVC ok

+++ src/main/java/cat/itacademy/s04/t02/n01/services/FruitaService.java
@@ -1,5 +1,6 @@
 package cat.itacademy.s04.t02.n01.services;
 
+import cat.itacademy.s04.t02.n01.exception.FruitNotFoundException;
 import cat.itacademy.s04.t02.n01.model.Fruita;
 import cat.itacademy.s04.t02.n01.repository.FruitaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,19 +13,28 @@ public class FruitaService {
     @Autowired
     private FruitaRepository repository;
 
-    public Fruita addFruita(Fruita fruita) {
+    public Fruita add(Fruita fruita) {
         return repository.save(fruita);
     }
 
-    public Fruita updateFruita(Fruita fruita) {
+    public Fruita update(Fruita fruita) {
+        if (fruita.getId() == 0 || !repository.existsById(fruita.getId())) {
+            throw new FruitNotFoundException("The id is null or the fruit no exist");
+        }
         return repository.save(fruita);
     }
 
-    public void deleteFruita(Long id) {
+    public void delete(int id) {
+        if (id == 0 || !repository.existsById(id)) {
+            throw new FruitNotFoundException("The id is null or the fruit no exist");
+        }
         repository.deleteById(id);
     }
 
-    public Fruita getOne(Long id) {
+    public Fruita getOne(int id) {
+        if (id == 0 || !repository.existsById(id)) {
+            throw new IllegalArgumentException("The id is null or the fruit no exist");
+        }
         return repository.findById(id).orElse(null);
     }
